requirements.txt
streamlit
python-dotenv
openai
langchain
langchain-core
langchain-community
chromadb
tiktoken
PyMuPDF
pypdf
pdfminer.six
sentence-transformers